enhance: use standard xxhash for int64 values
